register validacio 5 char minim nom, fix alguns validadors fotos... etc

+++ HubnRock/src/app/editar-solucio-esborrany/editar-solucio-esborrany.component.ts
@@ -96,6 +96,8 @@ export class EditarSolucioEsborranyComponent implements OnInit {
 
     },
     'videoSolucio': {
+      'pattern': 'No és un enllaç de Youtube vàlid'
+
     },
     'nomEquip': {
       'required': 'És un camp obligatori.',
@@ -150,6 +152,8 @@ export class EditarSolucioEsborranyComponent implements OnInit {
     this.currentTab = 0;
     this.radioValue = 'equip';
 
+    var regExp = /^.*(youtu.be\/|v\/|u\/\w\/|embed\/|watch\?v=|\&v=|\?v=)([^#\&\?]*).*/;
+
     this.solucioForm = this.fb.group({
       nomSolucio: ['', [Validators.required, Validators.maxLength(255), Validators.minLength(3)]],
       descripcioBreuSolucio: ['', [Validators.required, Validators.maxLength(280), Validators.minLength(3)]],  //max 280 chars en aquest
@@ -157,7 +161,7 @@ export class EditarSolucioEsborranyComponent implements OnInit {
       descripcioSolucio: ['', [Validators.required, Validators.maxLength(1000), Validators.minLength(3)]],  //textarea gran, 500 o 1000 va be?
       innovadoraSolucio: ['', [Validators.required, Validators.maxLength(1000), Validators.minLength(3)]],
       faseSolucio: ['', [Validators.required, Validators.maxLength(255), Validators.minLength(3)]],
-      videoSolucio: [''],
+      videoSolucio: ['', [Validators.pattern(regExp)]],
       pdf: [''],
       nomEquip: ['', [Validators.required, Validators.maxLength(255), Validators.minLength(3)]],
       membreArray: this.fb.array([
@@ -433,25 +437,26 @@ export class EditarSolucioEsborranyComponent implements OnInit {
 
   onPdfSelected(event) {
     if (event.target.files) {
+      let validSize = true;
       let totalSize = 0;
 
-      console.log(event.target.files[0].size, event.target.files)
       for (let index = 0; index < event.target.files.length; index++) {
         const element = event.target.files[index];
+
         totalSize += element.size
+
+        if (element.size > 1000000) {
+          validSize = false;
+
+        }
       }
 
-      if (totalSize < 15728640) {
+      if (validSize && totalSize < 15728640) {
         this.pdfArray = event.target.files
       } else {
+        alert('Un dels arxius pujats supera el límit de 1MB o la suma dels arxius és major a 15MB')
         this.pdfArray = null;
-        alert('Supera el límit de 15MB')
       }
-      // console.log(this.solucioForm.get('pdf').value)
-      // Array.from(this.pdfArray).forEach(file => {
-      //   console.log(file)
-      // });
-
     }
   }
 
@@ -514,44 +519,49 @@ export class EditarSolucioEsborranyComponent implements OnInit {
 
     }
     else {
-      this.formDone = true;
+      let confirmWindow = confirm('Està segur que vol enviar aquesta solució?')
 
-      if (this.formErrors.campsErronis) {
-        this.formErrors.campsErronis = '';
+      if (confirmWindow == true) {
+        this.formDone = true;
+
+        if (this.formErrors.campsErronis) {
+          this.formErrors.campsErronis = '';
+        }
+
+        let formData = this.appendRepte();
+
+        // this.subscriptionHttp1$ = this.httpClient.addSolucioValid(formData, this.idRepte)
+        //   .pipe(first())
+        //   .subscribe(
+        //     data => {
+        //       if (data.code == 1) {
+        //         this.idSolucioCreada = data.lastId;
+        //         this.success = true;
+        //         this.enviat = true;
+        //         console.log(data);
+
+        //         this.subscriptionHttp1$ = this.httpClient.deleteSolucio(this.idSolucio)
+        //           .pipe(first())
+        //           .subscribe();
+        //       }
+
+
+        //     });
+
+        this.subscriptionHttp1$ = this.httpClient.editSolucio(this.idSolucio, formData)
+          .pipe(first())
+          .subscribe(
+            data => {
+              if (data.code == 1) {
+                window.scrollTo(0, 0)
+
+                this.success = true;
+                this.enviat = true;
+              }
+            });
       }
-
-      let formData = this.appendRepte();
-
-      // this.subscriptionHttp1$ = this.httpClient.addSolucioValid(formData, this.idRepte)
-      //   .pipe(first())
-      //   .subscribe(
-      //     data => {
-      //       if (data.code == 1) {
-      //         this.idSolucioCreada = data.lastId;
-      //         this.success = true;
-      //         this.enviat = true;
-      //         console.log(data);
-
-      //         this.subscriptionHttp1$ = this.httpClient.deleteSolucio(this.idSolucio)
-      //           .pipe(first())
-      //           .subscribe();
-      //       }
-
-
-      //     });
-
-      this.subscriptionHttp1$ = this.httpClient.editSolucio(this.idSolucio, formData)
-        .pipe(first())
-        .subscribe(
-          data => {
-            if (data.code == 1) {
-              window.scrollTo(0, 0)
-
-              this.success = true;
-              this.enviat = true;
-            }
-          });
     }
+
 
 
   }
@@ -631,18 +641,23 @@ export class EditarSolucioEsborranyComponent implements OnInit {
   }
 
   deleteSolucio() {
-    this.subscriptionHttp1$ = this.httpClient.deleteSolucio(this.idSolucio)
-      .pipe(first())
-      .subscribe(
-        data => {
-          if (data.code == 1) {
-            window.scrollTo(0, 0)
+    let confirmWindow = confirm('Està segur que vol eliminar aquesta solució?')
 
-            this.success = true;
-            this.eliminat = true;
-          }
+    if (confirmWindow == true) {
+      this.subscriptionHttp1$ = this.httpClient.deleteSolucio(this.idSolucio)
+        .pipe(first())
+        .subscribe(
+          data => {
+            if (data.code == 1) {
+              window.scrollTo(0, 0)
 
-        });
+              this.success = true;
+              this.eliminat = true;
+            }
+
+          });
+    }
+
   }
 
   appendRepte(): FormData {
